fix(h2): drop unconvert lint warning on resetError ErrCode
frame.ErrCode() already returns uint32; the explicit uint32(...) cast
trips golangci-lint's unconvert check.

+++ h2client.go
@@ -701,7 +701,7 @@ func (hc *h2Conn) readLoop() {
 			idx := (frame.StreamID >> 1) % numSlots
 			chPtr := hc.streamSlots[idx].ch.Swap(nil)
 			if chPtr != nil {
-				*chPtr <- h2Response{err: resetError(uint32(frame.ErrCode()))}
+				*chPtr <- h2Response{err: resetError(frame.ErrCode())}
 			}
 
 		case frameSettings:
